requirements.txt
numpy
scipy
pytorch
firedrake